Everything related to a tenant in clusterDB needs to be in the same database
Its natural to organize data into collections (think of as a tables in sql land)

But in clustermgr and minion (soon), we use mongo watch notifications to figure out
what changed in the database and then react to it, like generating yamls for K8s.
As can be imagined, the order of notifications is important. But if the data is in
two different databases, there is no way to guarantee order of notifications in
mongo changeset. So we need to ensure that the "related" data (ie data needed to
be picked from multiple collections to create one logical data set) needs to come
from the same database if there are multiple collections - because then we can
guarantee the order of event processing is the same as order of event generation

So in this commit we make sure that the tenant information like the number of apods
and image used by the tenant etc.. comes from an NxtTenants collection that is in
the same database as the connector information - both will be in the Nxt-<cluster>-DB